extract types from tool store to separate file
