+++ client/src/stores/toolStoreTypes.ts
@@ -1,0 +1,52 @@
+export interface Tool {
+    model_class: string;
+    id: string;
+    name: string;
+    version: string;
+    description: string;
+    labels: string[];
+    edam_operations: string[];
+    edam_topics: string[];
+    hidden: "" | boolean;
+    is_workflow_compatible: boolean;
+    xrefs: string[];
+    config_file: string;
+    link: string;
+    min_width: number;
+    target: string;
+    panel_section_id: string;
+    panel_section_name: string | null;
+    form_style: string;
+    disabled?: boolean;
+}
+
+export interface ToolSection {
+    model_class: string;
+    id: string;
+    name: string;
+    title?: string;
+    version?: string;
+    description?: string;
+    links?: Record<string, string>;
+    tools?: (string | ToolSectionLabel)[];
+    elems?: (Tool | ToolSection)[];
+}
+
+export interface ToolSectionLabel {
+    model_class: string;
+    id: string;
+    text: string;
+    version?: string;
+    description?: string | null;
+    links?: Record<string, string> | null;
+}
+
+export interface FilterSettings {
+    [key: string]: string | undefined;
+    name?: string;
+    section?: string;
+    ontology?: string;
+    id?: string;
+    owner?: string;
+    help?: string;
+}
